added pathmagic
