+++ serpent-moltres/fuel-column/pathmagic.py
@@ -1,0 +1,3 @@
+import sys
+
+sys.path.insert(1, "/home/roberto/github/mhtgr350-benchmark")
